+++ internal/app/feed/export_feeds.go
@@ -2,8 +2,6 @@ package feed
 
 import (
 	"context"
-	"errors"
-	"sync"
 
 	"github.com/SafetyCulture/safetyculture-exporter/internal/app/api"
 	"github.com/SafetyCulture/safetyculture-exporter/internal/app/config"
@@ -12,69 +10,6 @@ import (
 )
 
 const maxConcurrentGoRoutines = 10
-
-// GetFeeds returns list of available SafetyCulture feeds
-func (e *ExporterApp) GetFeeds() []Feed {
-	return []Feed{
-		&InspectionFeed{
-			SkipIDs:       e.exportConfig.InspectionConfig.SkipIDs,
-			ModifiedAfter: e.exportConfig.ModifiedAfter,
-			TemplateIDs:   e.exportConfig.FilterByTemplateID,
-			Archived:      e.exportConfig.InspectionConfig.Archived,
-			Completed:     e.exportConfig.InspectionConfig.Completed,
-			Incremental:   e.exportConfig.Incremental,
-			Limit:         e.exportConfig.InspectionConfig.BatchLimit,
-			WebReportLink: e.exportConfig.InspectionConfig.WebReportLink,
-		},
-		&InspectionItemFeed{
-			SkipIDs:         e.exportConfig.InspectionConfig.SkipIDs,
-			ModifiedAfter:   e.exportConfig.ModifiedAfter,
-			TemplateIDs:     e.exportConfig.FilterByTemplateID,
-			Archived:        e.exportConfig.InspectionConfig.Archived,
-			Completed:       e.exportConfig.InspectionConfig.Completed,
-			IncludeInactive: e.exportConfig.InspectionConfig.IncludeInactiveItems,
-			Incremental:     e.exportConfig.Incremental,
-			Limit:           e.exportConfig.InspectionConfig.BatchLimit,
-			ExportMedia:     e.exportConfig.MediaConfig.Export,
-		},
-		&TemplateFeed{
-			Incremental: e.exportConfig.Incremental,
-		},
-		&TemplatePermissionFeed{
-			Incremental: e.exportConfig.Incremental,
-		},
-		&SiteFeed{
-			IncludeDeleted:       e.exportConfig.SiteConfig.IncludeDeleted,
-			IncludeFullHierarchy: e.exportConfig.SiteConfig.IncludeFullHierarchy,
-		},
-		&SiteMemberFeed{},
-		&UserFeed{},
-		&GroupFeed{},
-		&GroupUserFeed{},
-		&ScheduleFeed{
-			TemplateIDs: e.exportConfig.FilterByTemplateID,
-		},
-		&ScheduleAssigneeFeed{
-			TemplateIDs: e.exportConfig.FilterByTemplateID,
-		},
-		&ScheduleOccurrenceFeed{
-			TemplateIDs: e.exportConfig.FilterByTemplateID,
-		},
-		&ActionFeed{
-			ModifiedAfter: e.exportConfig.ModifiedAfter,
-			Incremental:   e.exportConfig.Incremental,
-			Limit:         e.exportConfig.ActionConfig.BatchLimit,
-		},
-		&ActionAssigneeFeed{
-			ModifiedAfter: e.exportConfig.ModifiedAfter,
-			Incremental:   e.exportConfig.Incremental,
-		},
-		&IssueFeed{
-			Incremental: false, //this was disabled on request. Issues API doesn't support modified After filters
-			Limit:       e.exportConfig.ActionConfig.BatchLimit,
-		},
-	}
-}
 
 // GetFeeds LEGACY returns list of all available data feeds
 func GetFeeds(v *viper.Viper) []Feed {
@@ -185,17 +120,6 @@ func GetSheqsyFeeds() []Feed {
 	}
 }
 
-// CreateSchemas generates schemas for the data feeds without fetching any data
-func (e *ExporterApp) CreateSchemas(exporter Exporter) error {
-	var lastErr error = nil
-	feeds := e.GetFeeds()
-	for _, feed := range feeds {
-		lastErr = feed.CreateSchema(exporter)
-	}
-
-	return lastErr
-}
-
 // WriteSchemas is used to print the schema of each feed to console output
 func WriteSchemas(v *viper.Viper, exporter *SchemaExporter) error {
 	logger := util.GetLogger()
@@ -210,103 +134,6 @@ func WriteSchemas(v *viper.Viper, exporter *SchemaExporter) error {
 	}
 
 	logger.Info("Writing schemas finished")
-	return nil
-}
-
-// ExportFeeds fetches all the feeds data from server and stores them in the format provided
-func (e *ExporterApp) ExportFeeds(exporter Exporter) error {
-	logger := util.GetLogger()
-	ctx := context.Background()
-
-	tables := e.exportConfig.FilterByTableName
-	tablesMap := map[string]bool{}
-	for _, table := range tables {
-		tablesMap[table] = true
-	}
-
-	var wg sync.WaitGroup
-	semaphore := make(chan int, maxConcurrentGoRoutines)
-
-	atLeastOneRun := false
-
-	// Run export for SafetyCulture data
-	if len(e.apiConfig.AccessToken) != 0 {
-		atLeastOneRun = true
-		logger.Info("exporting SafetyCulture data")
-
-		var feeds []Feed
-		for _, feed := range e.GetFeeds() {
-			if tablesMap[feed.Name()] || len(tables) == 0 {
-				feeds = append(feeds, feed)
-			}
-		}
-
-		resp, err := e.apiClient.WhoAmI(ctx)
-		util.Check(err, "failed to get details of the current user")
-
-		logger.Infof("Exporting data by user: %s %s", resp.Firstname, resp.Lastname)
-
-		if len(feeds) == 0 {
-			return errors.New("no tables selected")
-		}
-
-		for _, feed := range feeds {
-			semaphore <- 1
-			wg.Add(1)
-
-			go func(f Feed) {
-				logger.Infof(" ... queueing %s\n", f.Name())
-				defer wg.Done()
-				err := f.Export(ctx, e.apiClient, exporter, resp.OrganisationID)
-				util.Check(err, "failed to export")
-				<-semaphore
-			}(feed)
-		}
-	}
-
-	// Run export for SHEQSY data
-	if len(e.sheqsyApiConfig.UserName) != 0 {
-		atLeastOneRun = true
-		logger.Info("exporting SHEQSY data")
-
-		var feeds []Feed
-		for _, feed := range GetSheqsyFeeds() {
-			if tablesMap[feed.Name()] || len(tables) == 0 {
-				feeds = append(feeds, feed)
-			}
-		}
-
-		resp, err := e.sheqsyApiClient.GetSheqsyCompany(ctx, e.sheqsyApiConfig.CompanyID)
-		util.Check(err, "failed to get details of the current user")
-
-		logger.Infof("Exporting data for SHEQSY company: %s %s", resp.Name, resp.CompanyUID)
-
-		if len(feeds) == 0 {
-			return errors.New("no tables selected")
-		}
-
-		for _, feed := range feeds {
-			semaphore <- 1
-			wg.Add(1)
-
-			go func(f Feed) {
-				logger.Infof(" ... queueing %s\n", f.Name())
-				defer wg.Done()
-				err := f.Export(ctx, e.sheqsyApiClient, exporter, resp.CompanyUID)
-				util.Check(err, "failed to export")
-				<-semaphore
-			}(feed)
-		}
-	}
-
-	wg.Wait()
-
-	if !atLeastOneRun {
-		return errors.New("no API tokens provided")
-	}
-
-	logger.Info("Export finished")
-
 	return nil
 }
