+++ internal/app/inspections/inspections_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/SafetyCulture/safetyculture-exporter/cmd/safetyculture-exporter/cmd/export"
 	"github.com/stretchr/testify/require"
 
 	"github.com/SafetyCulture/safetyculture-exporter/internal/app/api"
@@ -52,11 +53,8 @@ func TestInspectionsExport(t *testing.T) {
 	exporterMock.On("SetLastModifiedAt", mock.Anything)
 	exporterMock.On("GetLastModifiedAt", mock.Anything).Return(nil)
 
-	inspectionClient := inspections.NewInspectionClient(
-		viperConfig,
-		apiClient,
-		exporterMock,
-	)
+	exporterAppCfg := export.MapViperConfigToConfigurationOptions(viperConfig)
+	inspectionClient := inspections.NewInspectionClient(exporterAppCfg, apiClient, exporterMock)
 	err := inspectionClient.Export(context.Background())
 	assert.NoError(t, err)
 }
@@ -71,11 +69,8 @@ func TestInspectionsExport_WhenSkipID(t *testing.T) {
 	exporterMock.On("SetLastModifiedAt", mock.Anything)
 	exporterMock.On("GetLastModifiedAt", mock.Anything).Return(nil)
 
-	inspectionClient := inspections.NewInspectionClient(
-		viperConfig,
-		apiClient,
-		exporterMock,
-	)
+	exporterAppCfg := export.MapViperConfigToConfigurationOptions(viperConfig)
+	inspectionClient := inspections.NewInspectionClient(exporterAppCfg, apiClient, exporterMock)
 	inspectionClient.(*inspections.Client).SkipIDs = []string{"audit_d7e2f55b95094bd48fac601850e1db63"}
 	err := inspectionClient.Export(context.Background())
 	assert.NoError(t, err)
@@ -91,11 +86,8 @@ func TestInspectionsExport_WhenModifiedAtIsNotNil(t *testing.T) {
 	exporterMock.On("SetLastModifiedAt", mock.Anything)
 	exporterMock.On("GetLastModifiedAt", mock.Anything).Return(&time.Time{})
 
-	inspectionClient := inspections.NewInspectionClient(
-		viperConfig,
-		apiClient,
-		exporterMock,
-	)
+	exporterAppCfg := export.MapViperConfigToConfigurationOptions(viperConfig)
+	inspectionClient := inspections.NewInspectionClient(exporterAppCfg, apiClient, exporterMock)
 	err := inspectionClient.Export(context.Background())
 	assert.NoError(t, err)
 }
@@ -104,7 +96,8 @@ func TestNewInspectionClient(t *testing.T) {
 	v := viper.GetViper()
 	require.NotNil(t, v)
 
-	res := inspections.NewInspectionClient(v, nil, nil)
+	exporterAppCfg := export.MapViperConfigToConfigurationOptions(v)
+	res := inspections.NewInspectionClient(exporterAppCfg, nil, nil)
 	require.NotNil(t, res)
 
 	client, ok := res.(*inspections.Client)
